pyproject.toml
[tool.poetry]
name = "todo"
version = "0.1.0"
description = ""
authors = ["ivan-karaberov"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.109.0"
uvicorn = {extras = ["standart"], version = "^0.27.0.post1"}
python-dotenv = "^1.0.1"
alembic = "^1.13.1"
black = "^24.1.1"
bcrypt = "^4.1.2"
pydantic-settings = "^2.1.0"
pyjwt = {extras = ["crypto"], version = "^2.8.0"}
python-multipart = "^0.0.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
